- (snmpd.c, context_parse.c, party_parse.c, system.[ch]):   - 64bit fix: get_myaddr() should return int not long.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@781 06827809-a52a-0410-b366-d66718629ded

+++ snmplib/context_parse.c
@@ -73,7 +73,7 @@ read_context_database(filename)
     oid proxyId[64];
     char name[64];	/* friendly name */
     struct contextEntry *cxp, *rp;
-    u_long myaddr;
+    u_int myaddr;
     int diff;
 
     fp = fopen(filename, "r");
